Database refactoring (#43)
* change name of directory to aircraft current handler

* Revert "change name of directory to aircraft current handler"

This reverts commit 39f6a6fba312ee4014454211bbce3305a79f6adc.

* change name of directory to aircraft current handler

* add func for dropping current_time_aircraft

use for testing CreateCurrentTimeAircraftTable

* remove return error in setupTestDB

* add method for creating history table

* add method for creating all current and future tables for database with test

* remove create current_time_aircraft and history_aircraft from interface Also made these table not exported

* regenerated database mock

* add method for inserting history data from current

teadown method in database_test now drops table while setup will create Add test for new method

* add method for retrieving aircraft by icao

regenerated mock database

* add handler for /aircraft/history endpoint with tests

* add check to handler tests if GeoJson standard is followed

* add history_aircraft columns to AddHistoryFromCurrent for better readability

* change misleading test name

* add new history handler to rest main

remove slash from endpoint constants
Fix StatusNoContent if check in aircraftHistoryHandler.go

* made aircraftHistory endpoint response follow GeoJsonStandard

Add new LineString structs to achieve this Change the name of previous structs to end with Point to reflect the difference Tests works as it should

* remove GeoJson prefix from GeoJson structs

* change GetAllCurrentAircraft and GetHistoryByIcao to return a list of structs that represents the rows Move the parsing part of these methods to separate methods in geoJSON.go

Moved the parsing logic from these methods to new methods in geoJSON.go. This change ensures that the methods in database.go are solely responsible for returning database models, improving cohesion and adhering to the Single Responsibility Principle.

* remove TODOs and improved TestAdsbDB_CreateAdsbTables

* move AircraftCurrentModel and AircraftHistoryModel to models directory under db

To keep the codebase organized and makes it easier to manage more database models where introduced. This way, all the code related to the database, including the models, is in one place, making it easier to understand and maintain.

* linting in aircraftCurrentHandler_test.go

* add tests for testing if Converting GeoJSON methods in geoJSON.go follow the GeoJSON schema

GeoJSON schema source: https://github.com/geojson/schema?tab=readme-ov-file

Using the overall GeoJSON schema in this implementation found at:
https://geojson.org/schema/GeoJSON.json

* remove if len(aircraft) == 0 from ConvertCurrentModelToGeoJson

That method responsibility is not about checking the data but converting it

* move db/models to global/models

Since models are accessed throughout the project and not only in db package

* add and rewritten old comments

* renamed adsb table names: from current_time_aircraft to aircraft_current and from history_aircraft to aircraft_history

This way there is a consistent naming, with a prefix of data type (aircraft_), and suffix of data type state (_current and _history)

* remove old comments from aircraftCurrentHandler.go

* implement REST service Contains the methods that are rest API relevant for accessing database logic

* change: go mod tidy

* create SBS service Contains the methods that are relevant to SBS stream processing and updating the database with new SBS data

* remove an error message when writing status no content

A 204 response is terminated by the first empty line after the header
fields because it cannot contain a message body:
https://datatracker.ietf.org/doc/html/rfc7231#section-6.3.5

* Add the history path to defaultHandler.go Add an error message if encoding json fails. Change 'CurrentAircraftPath' constant to 'AircraftCurrentPath'

* go mod tidy

* implement sbs service into reception main.go

* change User and Password variables to DbUser and DbPassword to not be confused with other similar variables that may come later

* add new error messages

* add ValidateURL to aircraftCurrentHandler.go with additional tests

---------

Co-authored-by: Andreas Follevaag Malde <andrefma@stud.ntnu.no>